Refactor Package for Job
