+++ src/main/java/at/aau/serg/websocketserver/game/GameLogic.java
@@ -3,9 +3,9 @@ package at.aau.serg.websocketserver.game;
 import at.aau.serg.websocketserver.player.Player;
 import at.aau.serg.websocketserver.board.BoardService;
 import at.aau.serg.websocketserver.board.Field;
-import at.aau.serg.websocketserver.session.Job;
-import at.aau.serg.websocketserver.session.JobRepository;
-import at.aau.serg.websocketserver.session.JobService;
+import at.aau.serg.websocketserver.session.job.Job;
+import at.aau.serg.websocketserver.session.job.JobRepository;
+import at.aau.serg.websocketserver.session.job.JobService;
 import lombok.Getter;
 import lombok.Setter;
