+++ src/main/java/at/aau/serg/websocketserver/websocket/broker/WebSocketBrokerController.java
@@ -10,8 +10,8 @@ import at.aau.serg.websocketserver.board.Field;
 import at.aau.serg.websocketserver.lobby.Lobby;
 import at.aau.serg.websocketserver.lobby.LobbyService;
 import at.aau.serg.websocketserver.messaging.dtos.*;
-import at.aau.serg.websocketserver.session.Job;
-import at.aau.serg.websocketserver.session.JobService;
+import at.aau.serg.websocketserver.session.job.Job;
+import at.aau.serg.websocketserver.session.job.JobService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
