+++ src/test/java/at/aau/serg/websocketserver/session/job/JobTest.java
@@ -1,4 +1,4 @@
-package at.aau.serg.websocketserver.session;
+package at.aau.serg.websocketserver.session.job;
 
 import org.junit.jupiter.api.Test;
